The guy can move now!!
Note: The program works only so that the green guy can move...so when u try out the game, make sure that one of the players chooses green colour. Still....a lot of work has to be done, but as of now, the game takes care of the condition of multiple sixes also, and gives an option to either move six steps forward, or remove another guy. It doesn't update the board properly when  two coins are there on the board. Press space bar to roll the dice. And note!! You have to click on the game board and then press space bar!!( I wasted a looot of time realising this :p)